Styled buttons on the NFC check in page & linted
