+++ src/NFCRegistration/UserDataReview.jsx
@@ -3,12 +3,11 @@ import { StyleSheet, View } from "react-native";
 import RegistrationPage from "./RegistrationPage";
 import RegistrationContext, {
   registrationPages,
-  reset,
   confirmUserData,
 } from "../contexts/RegistrationContext";
-import { colors, baseStyles, textStyles } from "../styleConfig";
+import { baseStyles, textStyles } from "../styleConfig";
 import BodyText from "../components/BodyText";
-import TextButton from "../components/TextButton";
+import RejectAcceptButtons from "./RejectAcceptButtons";
 
 /**
  * A detailed summary of a user's data
@@ -24,21 +23,11 @@ export default function UserDataReview() {
       {userData && (
         <View style={styles.page}>
           <BodyText style={styles.userData}>Is this {userData.name}?</BodyText>
-          <View style={styles.buttonContainer}>
-            <TextButton
-              onPress={() => dispatch({ type: reset })}
-              containerStyle={[styles.button, styles.cancelButton]}
-            >
-              No
-            </TextButton>
-            <TextButton
-              color={colors.primary}
-              onPress={() => dispatch({ type: confirmUserData })}
-              containerStyle={styles.button}
-            >
-              Yes
-            </TextButton>
-          </View>
+          <RejectAcceptButtons
+            acceptAction={() => dispatch({ type: confirmUserData })}
+            acceptText="Yes"
+            rejectText="No"
+          />
         </View>
       )}
     </RegistrationPage>
@@ -46,22 +35,6 @@ export default function UserDataReview() {
 }
 
 const styles = StyleSheet.create({
-  button: {
-    flex: 1,
-    marginHorizontal: baseStyles.spacing / 2,
-  },
-  buttonContainer: {
-    flexDirection: "row",
-    justifyContent: "space-around",
-    marginTop: 10,
-    padding: baseStyles.spacing,
-  },
-  cancelButton: {
-    backgroundColor: colors.cancel,
-    borderColor: colors.primary,
-    borderWidth: baseStyles.borderWidth,
-    overflow: "visible",
-  },
   page: {
     alignSelf: "stretch",
     flex: 1,
